requirements.txt
cryptography==45.0.5
docx2pdf==0.1.8
langchain==0.3.26
langchain-openai==0.3.27
openai==1.95.0
pdf2image==1.17.0
pdfplumber==0.11.7
pillow==11.3.0
poppler-utils==0.1.0
pymupdf==1.26.3
pypdf==5.7.0
python-docx==1.2.0
python-dotenv==1.1.1
streamlit==1.46.1
streamlit-pdf-viewer==0.0.26